Converting the old codebase
Changed the old GamePanel class to instead from handling all the game state and variables to moving it to a separate Board class which handles all the game logic. Implemented Observer methods for easier communication between the classes and also makes the different classes less intertwined with each other. This makes future implementation of features easier.